refactor: organize solo parsing functionality
